MCA2038
Good evening sir, Myself Shanjana S, Final Year, This is My Week 1 Assignment, The title is Text to Speech using python, As you may guess, gTTS stands for Google Text To Speech, it is a Python library to interface with Google Translate's text to speech API. It requires Internet connection and it's pretty easy to use.  The output of this program is what the text given is speech out. I use pycharm tool for this program.